display remove stream
